+++ libindy/src/api/wallet.rs
@@ -164,7 +164,7 @@ pub extern fn indy_register_wallet_storage(command_handle: i32,
 ///   "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                          Can be optional if storage supports default configuration.
 ///                          For 'default' storage type should be empty.
-///   "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+///   "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
 /// }
 ///
 /// #Returns
@@ -235,8 +235,8 @@ pub extern fn indy_create_wallet(command_handle: i32,
 ///       "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                              Can be optional if storage supports default configuration.
 ///                              For 'default' storage type should be empty.
-///       "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
-///       "rekey_simplified_security": Optional<bool> Use less secured but faster crypto algorithm for new master rekey derivation (false by default).
+///       "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+///       "rekey_simplified_security": optional<bool> Use less secured but faster crypto algorithm for new master rekey derivation (false by default).
 ///   }
 ///
 /// #Returns
@@ -288,7 +288,7 @@ pub extern fn indy_open_wallet(command_handle: i32,
 ///   {
 ///     "path": <string>, Path of the file that contains exported wallet content
 ///     "key": <string>, Passphrase used to derive export key
-///     "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for export key derivation (false by default).
+///     "simplified_security": optional<bool> Use less secured but faster crypto algorithm for export key derivation (false by default).
 ///   }
 ///
 /// #Returns
@@ -354,7 +354,7 @@ pub extern fn indy_export_wallet(command_handle: i32,
 ///   "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                          Can be optional if storage supports default configuration.
 ///                          For 'default' storage type should be empty.
-///   "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+///   "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
 /// }
 /// import_config: Import settings json.
 /// {
@@ -467,7 +467,7 @@ pub extern fn indy_close_wallet(command_handle: i32,
 ///   "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                          Can be optional if storage supports default configuration.
 ///                          For 'default' storage type should be empty.
-///   "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+///   "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
 /// }
 ///
 /// #Returns