+++ CHANGELOG.md
@@ -7,8 +7,10 @@
     * Updated `indy_build_pool_upgrade_request` API function to accept the additional parameter `package` that allow specify package to be upgraded.* Bugfixes
 * Added `pool restart` command in Indy CLI.
 * Updated Libindy CD pipeline to run iOS tests and to publish artifacts for Libindy and Libnullpay.
-* Bugfixes       
-  
+* Updated wallet `credentials` to accept the additional parameter `simplified_security`.
+  This parameter provides the ability to use less secured but faster crypto algorithm for master key derivation.
+* Bugfixes
+ 
 ## 1.6.1 bugfixes
 * Fix connection performance issue
 * Fix Android publishing